[0027] Prevent sending transport maps in player update object (will be replaced by a better fix later). Should fix the random teleports for this moment...
Fixed aura update (Fixes warrior/paladin client freeze on world login).

+++ src/game/SpellAuras.cpp
@@ -9184,7 +9184,7 @@ void SpellAuraHolder::BuildUpdatePacket(WorldPacket& data) const
     data << uint32(GetId());
 
     uint8 auraFlags = GetAuraFlags();
-    data << uint8(auraFlags);
+    data << uint16(auraFlags);
     data << uint8(GetAuraLevel());
 
     uint32 stackCount = m_procCharges ? m_procCharges * m_stackAmount : m_stackAmount;
